requirements.txt
anyio==4.4.0
autogen-agentchat~=0.2
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
numpy==1.26.4
openai==1.44.1
packaging==24.1
pandas==2.2.3
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
streamlit==1.40.0
gtts==2.2.3
streamlit_audiorecorder==0.0.6
pyttsx3==2.98
pydub==0.25.1
ffmpeg-python==0.2.0
openai-whisper==20240930